Revert "Indexing changes"
